Blind fix
...in --enable-online-update=mar code, after
7d8e94444d989d0ac4a4055b207726708e9ec0da "convert a<b?a:b to std::min(a,b)"

+++ onlineupdate/source/update/updater/bspatch.cxx
@@ -105,7 +105,7 @@ MBS_ApplyPatch(const MBSPatchHeader *header, FILE* patchFile,
     unsigned char *wb = buf;
     while (r)
     {
-        const size_t count = std::min(r, SSIZE_MAX);
+        const size_t count = std::min(r, size_t(SSIZE_MAX));
         size_t c = fread(wb, 1, count, patchFile);
         if (c != count)
         {
